Move cf from lib to dal
