+++ api-controllers/FabrikBaseController.js
@@ -7,7 +7,7 @@ const BaseController = require('../common/controllers/BaseController');
 const config = require('../common/config');
 const utils = require('../broker/lib/utils');
 const errors = require('../common/errors');
-const cf = require('../broker/lib/cf');
+const cf = require('../data-access-layer/cf');
 const bosh = require('../data-access-layer/bosh');
 const fabrik = require('../broker/lib/fabrik');
 const backupStore = require('../data-access-layer/iaas').backupStore;
