+++ api-controllers/ServiceFabrikApiController.js
@@ -15,7 +15,7 @@ const Unauthorized = errors.Unauthorized;
 const NotFound = errors.NotFound;
 const Gone = errors.Gone;
 const Timeout = errors.Timeout;
-const cf = require('../broker/lib/cf');
+const cf = require('../data-access-layer/cf');
 const Forbidden = errors.Forbidden;
 const BadRequest = errors.BadRequest;
 const Conflict = errors.Conflict;
