+++ broker/lib/fabrik/CfPlatformManager.js
@@ -6,7 +6,7 @@ const BasePlatformManager = require('./BasePlatformManager');
 const utils = require('../utils');
 const assert = require('assert');
 const errors = require('../../../common/errors');
-const cloudController = require('../cf').cloudController;
+const cloudController = require('../../../data-access-layer/cf').cloudController;
 const logger = require('../../../common/logger');
 const CONST = require('../../../common/constants');
 const config = require('../../../common/config');
