+++ broker/lib/fabrik/DirectorManager.js
@@ -7,7 +7,7 @@ const config = require('../../../common/config');
 const logger = require('../../../common/logger');
 const errors = require('../../../common/errors');
 const bosh = require('../../../data-access-layer/bosh');
-const cf = require('../cf');
+const cf = require('../../../data-access-layer/cf');
 const backupStore = require('../../../data-access-layer/iaas').backupStore;
 const utils = require('../utils');
 const Agent = require('./Agent');
