+++ broker/lib/fabrik/FabrikStatusPoller.js
@@ -14,7 +14,7 @@ const config = require('../../../common/config');
 const CONST = require('../../../common/constants');
 const EventLogInterceptor = require('../../../common/EventLogInterceptor');
 const DirectorManager = require('./DirectorManager');
-const serviceFabrikClient = require('../cf').serviceFabrikClient;
+const serviceFabrikClient = require('../../../data-access-layer/cf').serviceFabrikClient;
 
 class FabrikStatusPoller {
