+++ broker/lib/fabrik/ServiceFabrikOperation.js
@@ -7,7 +7,7 @@ const errors = require('../../../common/errors');
 const logger = require('../../../common/logger');
 const jwt = require('../jwt');
 const utils = require('../utils');
-const cf = require('../cf');
+const cf = require('../../../data-access-layer/cf');
 const DirectorManager = require('./DirectorManager');
 const cloudController = cf.cloudController;
 const Conflict = errors.Conflict;
