+++ broker/lib/index.js
@@ -4,7 +4,6 @@ exports.jwt = require('./jwt');
 exports.utils = require('./utils');
 exports.store = require('./store');
 exports.middleware = require('./middleware');
-exports.cf = require('./cf');
 exports.ScheduleManager = require('./jobs');
 exports.fabrik = require('./fabrik');
 exports.bootstrap = bootstrap;
