+++ broker/lib/jobs/BackupReaperJob.js
@@ -10,7 +10,7 @@ const BaseJob = require('./BaseJob');
 const errors = require('../../../common/errors');
 const ServiceInstanceNotFound = errors.ServiceInstanceNotFound;
 const NotFound = errors.NotFound;
-const cloudController = require('../cf').cloudController;
+const cloudController = require('../../../data-access-layer/cf').cloudController;
 const backupStoreForInstance = require('../../../data-access-layer/iaas').backupStore;
 const backupStoreForOob = require('../../../data-access-layer/iaas').backupStoreForOob;
 const ScheduleManager = require('./ScheduleManager');
