+++ broker/lib/jobs/BaseJob.js
@@ -8,7 +8,7 @@ const errors = require('../../../common/errors');
 const logger = require('../../../common/logger');
 const NotImplementedBySubclass = errors.NotImplementedBySubclass;
 const CONST = require('../../../common/constants');
-const serviceFabrikClient = require('../cf').serviceFabrikClient;
+const serviceFabrikClient = require('../../../data-access-layer/cf').serviceFabrikClient;
 const serviceBrokerClient = require('../utils/ServiceBrokerClient');
 const Repository = require('../../../common/db').Repository;
