+++ broker/lib/jobs/ScheduleBackupJob.js
@@ -10,7 +10,7 @@ const errors = require('../../../common/errors');
 const utils = require('../utils');
 const retry = utils.retry;
 const ServiceInstanceNotFound = errors.ServiceInstanceNotFound;
-const cloudController = require('../cf').cloudController;
+const cloudController = require('../../../data-access-layer/cf').cloudController;
 const backupStore = require('../../../data-access-layer/iaas').backupStore;
 const ScheduleManager = require('./ScheduleManager');
 //Above reference to schedulemanager leads to the below cyclic dependency:
