+++ broker/lib/jobs/ServiceInstanceUpdateJob.js
@@ -7,7 +7,7 @@ const config = require('../../../common/config');
 const BaseJob = require('./BaseJob');
 const errors = require('../../../common/errors');
 const utils = require('../utils');
-const cloudController = require('../cf').cloudController;
+const cloudController = require('../../../data-access-layer/cf').cloudController;
 const Fabrik = require('../fabrik');
 const catalog = require('../../../common/models/catalog');
 const CONST = require('../../../common/constants');
