+++ broker/lib/quota/QuotaManager.js
@@ -3,7 +3,7 @@
 const _ = require('lodash');
 const Promise = require('bluebird');
 const logger = require('../../../common/logger');
-const cf = require('../cf');
+const cf = require('../../../data-access-layer/cf');
 const catalog = require('../../../common/models/catalog');
 const config = require('../../../common/config');
 const CONST = require('../../../common/constants');
