+++ data-access-layer/cf/ResourceStream.js
@@ -4,8 +4,8 @@ const _ = require('lodash');
 const Promise = require('bluebird');
 const url = require('url');
 const Readable = require('stream').Readable;
-const logger = require('../../../common/logger');
-const utils = require('../utils');
+const logger = require('../../common/logger');
+const utils = require('../../broker/lib/utils');
 const parseUrl = _.partialRight(url.parse, true);
 const formatUrl = url.format;
