+++ data-access-layer/cf/TokenIssuer.js
@@ -2,9 +2,9 @@
 
 const Promise = require('bluebird');
 const TokenInfo = require('./TokenInfo');
-const config = require('../../../common/config');
-const errors = require('../../../common/errors');
-const logger = require('../../../common/logger');
+const config = require('../../common/config');
+const errors = require('../../common/errors');
+const logger = require('../../common/logger');
 const Unauthorized = errors.Unauthorized;
 
 class TokenIssuer {
