+++ test/test_broker/cf.CloudControllerClient.spec.js
@@ -3,7 +3,7 @@
 const Promise = require('bluebird');
 const _ = require('lodash');
 const formatUrl = require('url').format;
-const CloudControllerClient = require('../../broker/lib/cf/CloudControllerClient');
+const CloudControllerClient = require('../../data-access-layer/cf/CloudControllerClient');
 const errors = require('../../common/errors');
 const ServiceInstanceNotFound = errors.ServiceInstanceNotFound;
 const SecurityGroupNotFound = errors.SecurityGroupNotFound;
